feat: add default authorization question and pre-commit hooks
closes #36, #35
