+++ .changeset/smart-fans-end.md
@@ -1,0 +1,6 @@
+---
+"gboost": minor
+"gboost-common": minor
+---
+
+Add default authorization question and pre-commit hooks
