+++ docs/demos/notifications/package.json
@@ -8,16 +8,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@aws-amplify/ui-react": "^2.8.0",
-    "@fontsource/inter": "^4.5.4",
-    "aws-amplify": "^4.3.15",
+    "@aws-amplify/ui-react": "^2.10.1",
+    "@fontsource/inter": "^4.5.5",
+    "aws-amplify": "^4.3.16",
     "gboost-ui": "^0.8.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "vite": "^2.8.6"
   },
   "devDependencies": {
-    "@types/react": "^17.0.33",
+    "@types/react": "^17.0.40",
     "@types/react-dom": "^17.0.13",
     "@vitejs/plugin-react": "^1.0.7",
     "typescript": "^4.6.2"
