+++ docs/package.json
@@ -32,15 +32,15 @@
     "@docusaurus/module-type-aliases": "2.0.0-beta.16",
     "@docusaurus/plugin-content-docs": "^2.0.0-beta.17",
     "@docusaurus/theme-classic": "2.0.0-beta.16",
-    "@tsconfig/docusaurus": "^1.0.4",
-    "@types/react": "^17.0.39",
-    "eslint": "^8.10.0",
+    "@tsconfig/docusaurus": "^1.0.5",
+    "@types/react": "^17.0.40",
+    "eslint": "^8.11.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-config-react-app": "^7.0.0",
-    "eslint-define-config": "^1.2.5",
+    "eslint-define-config": "^1.3.0",
     "eslint-plugin-jest": "^25.7.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "prettier": "^2.5.1",
+    "prettier": "^2.6.0",
     "typescript": "^4.6.2"
   },
   "browserslist": {
