+++ package.json
@@ -22,7 +22,7 @@
     "@types/node": "^16.11.7",
     "husky": "^7.0.4",
     "license-checker": "^25.0.1",
-    "lint-staged": "^12.3.5",
+    "lint-staged": "^12.3.6",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.2"
   },
