+++ packages/gboost/_templates/repo/create/.husky/_/.gitignore.t
@@ -1,0 +1,5 @@
+---
+to: .husky/_/.gitignore
+---
+
+*
