+++ packages/gboost/_templates/repo/create/.husky/pre-commit.t
@@ -1,0 +1,8 @@
+---
+to: .husky/pre-commit
+---
+
+#!/bin/sh
+. "$(dirname "$0")/_/husky.sh"
+
+pnpm lint-staged
