+++ packages/gboost/_templates/repo/create/common/jest.config.cjs.t
@@ -1,5 +1,5 @@
 ---
-to: infra/jest.config.cjs
+to: common/jest.config.cjs
 ---
 
 /** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
