+++ packages/gboost/_templates/repo/create/common/src/groups.test.ts.t
@@ -1,0 +1,9 @@
+---
+to: common/src/groups.test.ts
+---
+
+import { groups } from "./groups.js";
+
+test("groups", () => {
+  expect(groups).toBeDefined();
+});
