+++ packages/gboost/_templates/repo/create/infra/src/front-end-stack.test.ts.t
@@ -1,0 +1,7 @@
+---
+to: infra/src/front-end-stack.test.ts
+---
+
+test("TODO", () => {
+  expect(true).toBe(true);
+});
