+++ packages/gboost/_templates/repo/create/ui/package.json.t
@@ -18,42 +18,48 @@ to: ui/package.json
     "test": "jest"
   },
   "dependencies": {
-    "@amcharts/amcharts5": "^5.1.4",
-    "@aws-amplify/api-graphql": "^2.2.21",
-    "@aws-amplify/auth": "^4.4.1",
-    "@aws-amplify/ui": "3.0.11",
-    "@aws-amplify/ui-react": "2.2.1",
-    "@fontsource/inter": "^4.5.4",
-    "@mantine/hooks": "^3.6.7",
+    "@amcharts/amcharts5": "^5.1.8",
+    "@aws-amplify/ui-react": "^2.10.1",
+    "@fontsource/inter": "^4.5.5",
+    "@mantine/hooks": "^3.6.14",
     "@radix-ui/colors": "^0.1.8",
-    "@stitches/react": "^1.2.6",
-    "aws-amplify": "^4.3.14",
+    "@stitches/react": "^1.2.7",
+    "aws-amplify": "^4.3.16",
     "gboost-common": "^0.3.3",
-    "gboost-ui": "^0.8.1",
+    "gboost-ui": "^0.6.2",
     "graphql": "^15.8.0",
     "graphql-tag": "^2.12.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
-    "react-router-dom": "^6.2.1"
+    "react-router-dom": "^6.2.2"
   },
   "devDependencies": {
-    "@types/jest": "^27.4.0",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@types/jest": "^27.4.1",
+    "@types/react": "^17.0.40",
+    "@types/react-dom": "^17.0.13",
     "@vitejs/plugin-react": "^1.2.0",
-    "eslint": "^8.9.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint": "^8.11.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-config-react-app": "^7.0.0",
-    "eslint-define-config": "^1.2.5",
+    "eslint-define-config": "^1.3.0",
     "eslint-plugin-jest": "^25.7.0",
     "eslint-plugin-prettier": "^4.0.0",
     "jest": "^27.5.1",
-    "prettier": "^2.5.1",
+    "prettier": "^2.6.0",
     "ts-jest": "^27.1.3",
-    "typescript": "^4.5.5",
-    "vite": "^2.8.3",
+    "typescript": "~4.5.5",
+    "vite": "^2.8.6",
     "vite-plugin-html": "^2.1.2",
-    "vite-tsconfig-paths": "^3.4.0"
+    "vite-tsconfig-paths": "^3.4.1"
+  },
+  "pnpm": {
+    "peerDependencyRules": {
+      "ignoreMissing": [
+        "@babel/plugin-syntax-flow",
+        "@babel/plugin-transform-react-jsx",
+        "react-native"
+      ]
+    }
   }
 }
