+++ packages/gboost/_templates/repo/create/ui/src/App.tsx.t
@@ -31,14 +31,12 @@ export function App() {
       <BreakpointsProvider media={config.media}>
         <NotificationsProvider>
           <Authenticator signUpAttributes={signUpAttributes} title={title}>
-            {({ user, signOut }) => (
+            {() => (
               <Layout
                 className={theme}
                 logoSrc="/favicon-32x32.png"
                 pages={pages}
-                signOut={signOut}
                 title={title}
-                user={user}
               />
             )}
           </Authenticator>
