+++ packages/gboost/_templates/repo/create/ui/src/main.tsx.t
@@ -2,8 +2,7 @@
 to: ui/src/main.tsx
 ---
 
-import Auth from "@aws-amplify/auth";
-import Api from "@aws-amplify/api-graphql";
+import { Auth, API } from "aws-amplify";
 import { render } from "react-dom";
 import { StrictMode } from "react";
 import { BrowserRouter } from "react-router-dom";
@@ -15,7 +14,7 @@ Auth.configure({
   userPoolWebClientId: import.meta.env.VITE_USER_POOL_CLIENT_ID,
 });
 
-Api.configure({
+API.configure({
   aws_appsync_graphqlEndpoint: import.meta.env.VITE_GQL_URL,
   aws_appsync_region: import.meta.env.VITE_AWS_REGION,
   aws_appsync_authenticationType: "AMAZON_COGNITO_USER_POOLS",
