+++ packages/gboost/_templates/repo/create/ui/src/pages/Dashboard/StackedColumnChart.tsx.t
@@ -1,5 +1,5 @@
 ---
-to: "<%= features.includes(`demoDashboard`) ? `ui/src/pages/Dashboard/StackedColumnChart.tsx` : null %>"
+to: "<%= features.includes(`Demo Dashboard`) ? `ui/src/pages/Dashboard/StackedColumnChart.tsx` : null %>"
 ---
 
 import { ReactElement, useLayoutEffect, useRef } from "react";
