+++ packages/gboost/_templates/repo/create/ui/src/theme.test.ts.t
@@ -1,0 +1,7 @@
+---
+to: ui/src/theme.test.ts
+---
+
+test("TODO", () => {
+  expect(true).toBe(true);
+});
