+++ packages/gboost/_templates/repo/create/ui/tsconfig.json.t
@@ -11,7 +11,6 @@ to: ui/tsconfig.json
     "esModuleInterop": false,
     "allowSyntheticDefaultImports": true,
     "resolveJsonModule": true,
-    "isolatedModules": true,
     "noEmit": true,
     "jsx": "react-jsx",
     "baseUrl": ".",
