+++ packages/gboost/_templates/repo/create/utils/lintStaged.js.t
@@ -1,0 +1,16 @@
+---
+to: utils/lintStaged.js
+---
+
+import { ESLint } from "eslint";
+
+export const removeIgnoredFiles = async (files) => {
+  const eslint = new ESLint();
+  const isIgnored = await Promise.all(
+    files.map((file) => {
+      return eslint.isPathIgnored(file);
+    })
+  );
+  const filteredFiles = files.filter((_, i) => !isIgnored[i]);
+  return filteredFiles.join(" ");
+};
