+++ packages/gboost/package.json
@@ -3,7 +3,7 @@
   "version": "0.3.10",
   "description": "Green Boost Command Line Interface",
   "bin": {
-    "gboost": "lib/index.js"
+    "gboost": "src/index.ts"
   },
   "main": "./lib/index.js",
   "module": "./lib/index.js",
@@ -15,13 +15,18 @@
   "engines": {
     "node": ">=16"
   },
+  "publishConfig": {
+    "bin": {
+      "gboost": "lib/index.js"
+    }
+  },
   "scripts": {
     "build": "rimraf lib && tsc",
     "lint": "eslint \"src/**/*.ts\"",
     "test": "jest"
   },
   "dependencies": {
-    "chalk": "^5.0.0",
+    "chalk": "^5.0.1",
     "enquirer": "^2.3.6",
     "hygen": "^6.2.0",
     "minimist": "^1.2.5"
@@ -31,15 +36,15 @@
     "@types/minimist": "^1.2.2",
     "@types/node": "^16.11.26",
     "@types/prompts": "^2.4.0",
-    "@typescript-eslint/eslint-plugin": "^5.14.0",
-    "@typescript-eslint/parser": "^5.14.0",
-    "eslint": "^8.10.0",
+    "@typescript-eslint/eslint-plugin": "^5.15.0",
+    "@typescript-eslint/parser": "^5.15.0",
+    "eslint": "^8.11.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-define-config": "^1.2.5",
+    "eslint-define-config": "^1.3.0",
     "eslint-plugin-jest": "^25.7.0",
     "eslint-plugin-prettier": "^4.0.0",
     "jest": "^27.5.1",
-    "prettier": "^2.5.1",
+    "prettier": "^2.6.0",
     "rimraf": "^3.0.2",
     "ts-jest": "^27.1.3",
     "ts-node": "^10.7.0",
@@ -66,5 +71,13 @@
     "url": "https://github.com/awslabs/green-boost/issues",
     "directory": "packages/gboost"
   },
-  "homepage": "https://awslabs.github.io/green-boost"
+  "homepage": "https://awslabs.github.io/green-boost",
+  "pnpm": {
+    "peerDependencyRules": {
+      "ignoreMissing": [
+        "eslint",
+        "prettier"
+      ]
+    }
+  }
 }
